fix(daemon,web): report analytics for daemon-started Runs and mid-write Labs toggles
Daemon-started OD Next continuations and background Runs now go
through the shared analytics lifecycle so every physical Run reports
run_created and run_finished with lineage. Labs toggle no longer
drops telemetry or strands the save pill when the section unmounts
mid-write.

Fixes #7485
